backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Login==0.6.3
Werkzeug==2.3.7
SQLAlchemy==2.0.23
python-dotenv==1.0.0
bcrypt==4.1.1
Pillow==10.1.0
requests==2.31.0
spotipy==2.23.0
mutagen==1.47.0
email-validator==2.1.0
python-multipart==0.0.6
psycopg2-binary==2.9.7 